chore: Update Launcher component to conditionally show templates based on the number of projects
The Launcher component in Launcher.tsx has been updated to conditionally show templates based on the number of projects. The 'showTemplates' property of the 'store.uiState' object is now set to 'true' only if the 'store.projects' array is not null and its length is less than 10. This change improves the user experience by dynamically displaying templates when there are fewer than 10 projects available.